Changed 101?

+++ 0x03-more_functions_nested_loops/101-print_number.c
@@ -1,6 +1,5 @@
 #include "holberton.h"
 
-
 /**
  * mypow - Quick pow implementation.
  * Description: Multiples number by itself a number of times
@@ -45,7 +44,7 @@ int _abs(int n)
  */
 void print_number(int n)
 {
-	int digits, stop, i, num;
+	long digits, stop, i, num;
 
 	i = 1;
 	digits = 0;
